+++ R/parse_plumber_file.R
@@ -14,7 +14,7 @@
 #' * `route` The main route handling requests according to the parsed file, as a
 #'   named list of length one
 #' * `header_route` The route to be attached to header events (fires before the
-#'   body has been recieved and can be used to prematurely reject requests based
+#'   body has been received and can be used to prematurely reject requests based
 #'   on their headers), as a named list of length one
 #' * `asset_routes` All the asset routes created by `@static` blocks as a named
 #'   list
@@ -116,6 +116,8 @@ parse_block <- function(
     parse_asset_block(call, tags, values, env, file_dir)
   } else if (block_has_tags(block, "statics")) {
     parse_static_block(call, tags, values, env, file_dir)
+  } else if (block_has_tags(block, "authenticator")) {
+    parse_authenticator_block(call, tags, values, env)
   } else if (block_has_tags(block, "message")) {
     parse_message_block(call, tags, values, env)
   } else if (block_has_tags(block, "then")) {
@@ -251,7 +253,8 @@ parse_handler_block <- function(call, tags, values, env) {
       download = download,
       doc = doc,
       async = async,
-      header = any(tags == "header")
+      header = any(tags == "header"),
+      auth = parse_auth_tags(tags, values)
     ),
     class = "plumber2_handler_block"
   )
@@ -263,12 +266,6 @@ parse_static_block <- function(call, tags, values, env, file_dir) {
   }
   if (!(is.null(call) || identical(call, list()))) {
     cli::cli_warn("Expression ignored for {.field @statics} blocks")
-  }
-  extra_tags <- setdiff(tags, c("statics", "except", "backref"))
-  if (length(extra_tags) != 0) {
-    cli::cli_warn(
-      "Ignoring {.field {paste0('@', extra_tags)}} tag{?s} when parsing {.field @statics} tag"
-    )
   }
   mapping <- trimws(strsplit(
     values[[which(tags == "statics")]],
@@ -300,12 +297,6 @@ parse_asset_block <- function(call, tags, values, env, file_dir) {
   if (!(is.null(call) || identical(call, list()))) {
     cli::cli_warn("Expression ignored for {.field @assets} blocks")
   }
-  extra_tags <- setdiff(tags, c("assets", "backref"))
-  if (length(extra_tags) != 0) {
-    cli::cli_warn(
-      "Ignoring {.field {paste0('@', extra_tags)}} tag{?s} when parsing {.field @assets} tag"
-    )
-  }
   mapping <- trimws(strsplit(
     values[[which(tags == "assets")]],
     " ",
@@ -315,13 +306,32 @@ parse_asset_block <- function(call, tags, values, env, file_dir) {
     mapping <- c(mapping, "/")
   }
   mapping[1] <- fs::path_abs(mapping[1], file_dir)
+  path <- mapping[2]
+  if (grepl("/$", path)) {
+    path <- paste0(path, "*")
+  }
   structure(
     list(
       route = routr::resource_route(!!mapping[2] := mapping[1]),
       header = FALSE,
-      endpoints = list(list(method = "get", path = mapping[2]))
+      endpoints = list(list(method = "get", path = path)),
+      auth = parse_auth_tags(tags, values)
     ),
     class = "plumber2_route_block"
+  )
+}
+
+parse_authenticator_block <- function(call, tags, values, env) {
+  if (!is.function(call) || fireproof::is_auth(call)) {
+    stop_input_type(call, "an {.cls Auth} subclass object or a function")
+  }
+  name <- trimws(values[[which(tags == "authenticator")[1]]])
+  structure(
+    list(
+      auth = call,
+      name = name
+    ),
+    class = "plumber2_authenticator_block"
   )
 }
 
@@ -394,7 +404,8 @@ parse_shiny_block <- function(call, tags, values, env) {
     list(
       shiny_app = call,
       path = values[[which(tags == "shiny")]],
-      except = unlist(values[except])
+      except = unlist(values[except]),
+      auth = parse_auth_tags(tags, values)
     ),
     class = "plumber2_proxy_block"
   )
@@ -422,7 +433,8 @@ parse_forward_block <- function(call, tags, values, env) {
     list(
       path = vapply(res, `[[`, character(1), "path"),
       url = vapply(res, `[[`, character(1), "url"),
-      except = unlist(values[except])
+      except = unlist(values[except]),
+      auth = parse_auth_tags(tags, values)
     ),
     class = "plumber2_proxy_block"
   )
@@ -459,7 +471,8 @@ parse_report_block <- function(call, tags, values, env, file_dir) {
       path = x,
       report = call,
       doc = doc,
-      endpoints = endpoints
+      endpoints = endpoints,
+      auth = parse_auth_tags(tags, values)
     ),
     class = "plumber2_report_block"
   )
@@ -516,14 +529,18 @@ apply_plumber2_block.plumber2_proxy_block <- function(
     api$add_shiny(
       paste0(root, block$path),
       block$shiny_app,
-      except = block$except
+      except = block$except,
+      auth_flow = !!block$auth$flow,
+      auth_scope = block$auth$scope
     )
   } else if (!is.null(block$url)) {
     for (i in seq_along(block$path)) {
       api$forward(
         paste0(root, block$path[i]),
         block$url[i],
-        except = block$except
+        except = block$except,
+        auth_flow = !!block$auth$flow,
+        auth_scope = block$auth$scope
       )
     }
   }
@@ -580,6 +597,19 @@ apply_plumber2_block.plumber2_route_block <- function(
   api$add_route(route_name, block$route, block$header, root = root)
   if (!is.null(block$doc)) {
     api$add_api_doc(block$doc)
+  }
+  if (!is.null(block$auth)) {
+    for (i in seq_along(block$endpoints)) {
+      for (path in block$endpoints[[i]]$path) {
+        api$add_authentication(
+          method = block$endpoints[[i]]$method,
+          path = paste0(root, path),
+          auth_flow = !!block$auth$flow,
+          auth_scope = block$auth$scope,
+          add_doc = FALSE
+        )
+      }
+    }
   }
   api
 }
@@ -644,6 +674,8 @@ apply_plumber2_block.plumber2_handler_block <- function(
       serializers = block$serializers,
       parsers = block$parsers,
       use_strict_serialize = block$use_strict_serializer,
+      auth_flow = !!block$auth$flow,
+      auth_scope = block$auth$scope,
       download = block$download,
       async = block$async,
       then = block$then,
@@ -662,7 +694,27 @@ apply_plumber2_block.plumber2_report_block <- function(
   root,
   ...
 ) {
-  api$add_report(block$path, block$report, doc = block$doc)
+  api$add_report(
+    block$path,
+    block$report,
+    doc = block$doc,
+    auth_flow = !!block$auth$flow,
+    auth_scope = block$auth$scope
+  )
+  api
+}
+#' @export
+apply_plumber2_block.plumber2_authenticator_block <- function(
+  block,
+  api,
+  route_name,
+  root,
+  ...
+) {
+  api$add_authenticator(
+    block$auth,
+    block$name
+  )
   api
 }
 #' @export
@@ -730,4 +782,26 @@ apply_plumber2_block.plumber2_rip_block <- function(
     }
   }
   api
+}
+
+
+# Helpers ----------------------------------------------------------------
+
+parse_auth_tags <- function(tags, values) {
+  if ("auth" %in% tags) {
+    flow <- trimws(values[[which(tags == "auth")[1]]])
+    flow <- parse_quo(flow, empty_env())
+    if ("authScope" %||% tags) {
+      scope <- unlist(values[tags == "authScope"])
+      scope <- unlist(trimws(unlist(strsplit(scope, ","))))
+    } else {
+      scope <- NULL
+    }
+    list(
+      flow = flow,
+      scope = scope
+    )
+  } else {
+    NULL
+  }
 }